+++ frontend/src/features/stream-control/useStreamControl.js
@@ -1,11 +1,26 @@
 import React from 'react';
 import {useNear} from '../near-connect/useNear';
 
+const STREAM_CONTROL_ID_NOT_PROVIDED = new Error(
+  'Provide stream id to use stream methods',
+);
+
 export function useStreamControl(streamId) {
+  const ensureMethodHasStreamId = () => {
+    if (!streamId) throw STREAM_CONTROL_ID_NOT_PROVIDED;
+  };
   const near = useNear();
   const [loading, setLoading] = React.useState(false);
 
+  async function updateAllAndWithdraw() {
+    console.log('updating all account');
+
+    await near.contractApi.updateAccount({});
+    console.log('update completed');
+  }
+
   async function deposit({deposit}) {
+    ensureMethodHasStreamId();
     console.log('depositing', streamId);
     await near.contractApi.depositStream({
       streamId,
@@ -15,6 +30,7 @@ export function useStreamControl(streamId) {
   }
 
   async function pause() {
+    ensureMethodHasStreamId();
     console.log('pausing', streamId);
     const res = await near.contractApi.pauseStream({
       streamId,
@@ -25,6 +41,7 @@ export function useStreamControl(streamId) {
   }
 
   async function restart() {
+    ensureMethodHasStreamId();
     console.log('restarting', streamId);
     const res = await near.contractApi.startStream({streamId: streamId});
     console.log('restarting res', res);
@@ -33,6 +50,7 @@ export function useStreamControl(streamId) {
   }
 
   async function stop() {
+    ensureMethodHasStreamId();
     console.log('Stop called stopping', streamId);
     const res = await near.contractApi.stopStream({streamId: streamId});
     console.log('stopping res', res);
@@ -40,21 +58,22 @@ export function useStreamControl(streamId) {
   }
 
   function wrapped(fn) {
-    return (...args) => {
-      console.log('wrapped', ...args);
+    return async (...args) => {
       if (loading) return;
 
       setLoading(true);
 
       try {
-        return fn(...args);
+        return await fn(...args);
       } finally {
         setLoading(false);
       }
     };
   }
+
   return {
     loading,
+    updateAllAndWithdraw: wrapped(updateAllAndWithdraw),
     pause: wrapped(pause),
     restart: wrapped(restart),
     stop: wrapped(stop),